First attempt at sharing intrinsics
Only handles intrinsics which don't produce diagnostics and don't "mark".
