+++ src/ILLink.Shared/TypeSystemProxy/IMemberProxy.cs
@@ -1,0 +1,10 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace ILLink.Shared.TypeSystemProxy
+{
+	interface IMemberProxy
+	{
+		public string Name { get; }
+	}
+}
