+++ test/ILLink.RoslynAnalyzer.Tests/DataFlowTests.cs
@@ -98,6 +98,12 @@ namespace ILLink.RoslynAnalyzer.Tests
 		}
 
 		[Fact]
+		public Task GetTypeInfoDataFlow ()
+		{
+			return RunTest (nameof (GetTypeInfoDataFlow));
+		}
+
+		[Fact]
 		public Task IReflectDataflow ()
 		{
 			return RunTest (nameof (IReflectDataflow));
